+++ cellarium/ml/transforms/__init__.py
@@ -1,6 +1,10 @@
 # Copyright Contributors to the Cellarium project.
 # SPDX-License-Identifier: BSD-3-Clause
 
-from .transforms import DivideByScale, Filter, Log1p, NormalizeTotal, ZScore
+from .divide_by_scale import DivideByScale
+from .filter import Filter
+from .log1p import Log1p
+from .normalize_total import NormalizeTotal
+from .z_score import ZScore
 
 __all__ = ["DivideByScale", "Filter", "Log1p", "NormalizeTotal", "ZScore"]
