+++ cellarium/ml/transforms/log1p.py
@@ -1,0 +1,32 @@
+# Copyright Contributors to the Cellarium project.
+# SPDX-License-Identifier: BSD-3-Clause
+
+
+import torch
+from torch import nn
+
+
+class Log1p(nn.Module):
+    """
+    Log1p transform gene counts.
+
+    .. math::
+
+        y_{ng} = \\log(1 + x_{ng})
+    """
+
+    def __init__(self) -> None:
+        super().__init__()
+
+    def forward(self, x_ng: torch.Tensor) -> torch.Tensor:
+        """
+        Args:
+            x_ng: Gene counts.
+
+        Returns:
+            Log1p transformed gene counts.
+        """
+        return torch.log1p(x_ng)
+
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}()"
