+++ docs/source/lr_schedulers.rst
@@ -1,0 +1,6 @@
+LR Schedulers
+=============
+
+.. automodule:: cellarium.ml.lr_schedulers
+   :members:
+   :show-inheritance:
