--- docs/source/optim.rst
@@ -1,6 +1,0 @@
-Optimization
-============
-
-.. automodule:: cellarium.ml.optim
-   :members:
-   :show-inheritance:
